moved the states data to a new component and updates all of the import statements for the form. Updated the information in the settings.example file
