+++ src/Components/StaticData.js
@@ -1,0 +1,6 @@
+//just a place to keep an array of state abbreviations to use in multiple forms
+export const states = ['AL','AK','AS','AZ','AR','CA','CO','CT','DE','DC','FM','FL','GA',
+'GU','HI','ID','IL','IN','IA','KS','KY','LA','ME','MH','MD','MA',
+'MI','MN','MS','MO','MT','NE','NV','NH','NJ','NM','NY','NC','ND',
+'MP','OH','OK','OR','PW','PA','PR','RI','SC','SD','TN','TX','UT',
+'VT','VI','VA','WA','WV','WI','WY']
