+++ src/Components/members/EditMember.js
@@ -1,7 +1,8 @@
 import { useContext, useEffect } from "react";
 import { MemberContext } from "./MemberProvider";
 import React, { useState } from "react"
-import { cloudinaryKeys, states } from "../settings"
+import { cloudinaryKeys } from "../settings"
+import { states } from "../StaticData"
 import { GroupContext } from "../groups/GroupProvider";
 import { Image, Placeholder } from 'cloudinary-react'
 import './EditMember.css'
